Update studio10LoopsAandC
Sierra showed me I do not need "-1" on the ".length" because I want it to stop at the length number. The ".length-1" is when we want to stop counting before the final count in the case where that passes the index. In this case, .length is sufficient because that is how far we want to go