Update candidateTestingPart3.js
At this stage, the questions are asked twice but the values aren't stored when I print the answer total. I will have to test the individual answers first and check their values to make sure it's not just the counter messing up. Undefined prints which makes me think the individual candidate answers are not storing the user inputs as values or losing the value/breaking when changing into an array.