Update objectStudio01.js
this is how far we got with mob programming in our group; we bounced some solid ideas off one another. Thanks to Adam, Kirsten, Mitch and Nhu, Sierra, and Max as well for coaching/mentoring us through this.

I suggested a nested 'for in' loop within a 'for' loop to compare the arrays, but Max suggested we loop through the crew to check for ids in the array and that would be less times that we have to do the loop. It gets into big O notation.