Update assignment2ScrabbleScorer.js
repl froze. I was trying a couple of things here. "return i" isn't giving me any output but the for loop works when I test with the "consolelog(i)" statement. I'll refresh and run again to make sure it still works. Thinking of commenting out the 'console.log' test code and leaving it for now.