Update assignment2ScrabbleScorer.js
This works but iterating through the 'word' parameter given is not enough because in this case, 'money' is shorter than the array of 'vowels' by 1. I should either iterate through the vowels checking the word or make the condition of the loop '(i < vowels.length)' so it goes through all the vowels.