Update assignment2ScrabbleScorer.js
I changed the 'score_function' key value in all the objects to the actual functions the objects represent. That's what we learned in our lessons, objects being able to hold functions, so it makes sense. Also, 2 descriptions in the same object doesn't make sense to me. I have a save state to grab that code back if need be.