Update assignment2ScrabbleScorer.js
Idk why I have to do this commit for the 3rd time but I believe these are the changes that made the program start acting funny right before I had to sign off. Github keeps showing this assignment as the one that was updated 20 mins ago even though it was only a few mins ago. Mercury is not even in gatorade