Update assignment2ScrabbleScorer.js
I think I'm still getting closer. I commented out the console log test on line 211 and the return statement on 218 seems like it's doing the correct thing. Now I just have to see why the value from the object in parameter isn't passing to the newKey but that may require an additional line of code that 209 can't take care of on it's own.