Update assignment2ScrabbleScorer.js
getting errors again but almost done. Trying to accomplish this part:

"3. Once you've defined newPointStructure, use it to finish writing the scrabbleScore() function and then replace the oldScrabbleScorer() function in scoringAlgorithms with this new function.

Tip
oldScrabbleScorer() uses oldPointStructure and returns a score for each letter in a word. You'll want to write scrabbleScore() to use newPointStructure and return a cumulative score for the whole word entered."

I have to change not just the array that contains all the objects but the oldScrabbleScorerObject itself, or create a new object and map to it in the array and while loop that restricts the user input to just the options available.